atacama: support Literals and Sets out of the box
Additionally, provide a basic TypeHandler lookup option that allows
downstream users to easily hook into Field generation with whatever
logic they want to add. This is _essentially_ the same as cattrs'
register_structure_hook and register_unstructure_hook, only cattrs
allows them to be split and marshmallow does not.

The Literal handling is built on top of this functionality. The Set
functionality has been built in for the time being, though arguably it
could be built using this same feature.

GitOrigin-RevId: 9dcf691ff4fad9bff1a01451ee7b9aa1f87e7cb0