atacama: handle uncollapsed optional unions
Generally, there is a single argument to an `Optional` type. The
existing atacama code, inherited from desert, assumed this would hold
for all circumstances.

However, it _appears_ that `Union`s of `Literal`s embedded in an
`Optional` actually get unwrapped by `typing_inspect` into a tuple of
arguments, rather than as either A) a `Union` or B) a collapsed
`Literal`. This results in a rather odd runtime type that were weren't
handling. Technically, a Union of Literals is identical to a single
Literal with all of the arguments concatenated together, and I am
surprised that `typing_inspect` does not fully unwrap and collapse the
Union of Literals in this special case (it does seem to collapse
Literals in the non-Optional case).

It would likely be possible to do this coalescing ourselves, but I'm
worried there may be other (non-Literal) cases that typing_inspect
doesn't handle well. I think it's a more general solution to instead
fall back to using the existing `Union` code any time we find multiple
non-NoneType arguments to an Optional type.

GitOrigin-RevId: 1a0d32fe7d1fd1aa970ef39d4c6ede8b1c6cda76