core fretry: fix longstanding bug in retry_sleep decorator
The bug was that sleep would only call its input iterable once,
and in most cases that iterable was not re-entrant, so the second time a failure was
encountered, there would be no retry at all. The tests only tested the first time
through...

Technically this is a backward-incompatible change, but in practice all callers used
retry_sleep and expo the same way, so changing the code in both places meant that no
callers needed to change.

GitOrigin-RevId: 7c0427bd372afc0d19cbb15232172ebf8a423b19