new blob_store_lock implementation
intended to help coordinate across multiple mops orchestrators when
they happen to call the same function with the same arguments.

In order to make this work, we'll need to introduce a new daemon thread
both for orchestrators and for remote invokers, so that they can
keep the lock 'fresh' for as long as something is actually running.

This will never be a perfect solution, because orchestrators can die
and remotely-invoked functions can get stuck in queues before Python
gets a chance to run. Also, this algorithm has known failure modes.

However, it seems at least reasonably plausible that we'll manage to make this work for
users without confusing them in 99% of cases.

-

GitOrigin-RevId: e30137fe93d50e18beb92aa8e0e0d6aff7493e7e