mops: k8s job log watching is now lazy.
use_runner has a simpler mechanism for disabling it, and it allows for
mops to be used recursively via manifests.

Some tweaks to k8s uncertain future usage so that it can be used by a batching shim.

cleaned up Shim types now that we're allowing Shims to return a Future.

GitOrigin-RevId: 0c6fe7c13565e2fc9006ee6686b4497abab7a389