A `fork` made our client conflate, Mixing paths at a furious rate. An `assert` stopped the crime, But we've fixed it sublime, So this check we now eliminate.
---

once, we had a bug where some state inside the azure storage third party Python SDK was getting confused because of forked processes. at first we couldn't figure out what was happening, except that it _seemed_ that we were somehow asking azure to download one file, and then it was fetching file properties for a completely different file.

so we added some assertions to help us identify this earlier on, rather than after the download, when it was quite confusing.

we later actually figured out that it was related to fork() on Linux, and we figured out how to avoid reusing the same client after fork.

now we still have these assertions, and they occasionally in edge cases cause issues because we're checking two paths against each other, and one of them has had a trailing slash stripped. since they're no longer going to help us track down any issue, it's not worth fixing the trailing slash issue; i'm just going to remove them entirely.

GitOrigin-RevId: 34bebf737282d464710b682ef9b6a16691c71c52